% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_easy_button.R
\name{map_easy_button}
\alias{map_easy_button}
\title{Add buttons on map (topleft corner)}
\usage{
map_easy_button(id, icon_file, label)
}
\arguments{
\item{id}{Id for the button (use session$ns("id") server-side)}

\item{icon_file}{Name of the icon, including the extension (.png or whatever), see inst/icons for list of available icons}

\item{label}{Label on hover (optional)}
}
\description{
See example apps for usage.
}
