% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoBaseMap.R
\name{shintoBaseMap}
\alias{shintoBaseMap}
\title{Makes a leaflet map for use in shintoMapModule}
\usage{
shintoBaseMap(
  set_view = NULL,
  fullscreen_option = FALSE,
  layers_button_position = "topright",
  ...
)
}
\description{
Makes a leaflet map for use in shintoMapModule
}
