% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoMapModule.R
\name{shintoMapModule}
\alias{shintoMapModule}
\title{Shinto Map server module}
\usage{
shintoMapModule(
  input,
  output,
  session,
  base_map,
  render_not_visible = TRUE,
  border = shiny::reactive(NULL),
  border_weight = 2,
  border_color = "#4f4f4f",
  auto_recenter = FALSE,
  label_function = function(data, ...) NULL,
  label_params = list(),
  color_default = "#023e8a",
  color_outline = "#FFFFFF",
  toggle_reload = shiny::reactive(TRUE),
  proxy = TRUE,
  layers = list()
)
}
\arguments{
\item{input}{shiny input, do not set}

\item{output}{shiny output, do not set}

\item{session}{shiny session, do not set}

\item{base_map}{Static map (with tiles and view setting) made with \code{\link[=shintoBaseMap]{shintoBaseMap()}}}

\item{render_not_visible}{If TRUE, renders the map even when not visible. Works in tabsetpanels (tab_box),
but does not work in modals! For modals, set render_not_visible = FALSE (map should be rendered only after
the modal is opened)}

\item{border}{Reactive sf dataframe with a polygon used as a border}

\item{border_weight}{Thickness of border polygon}

\item{border_color}{Color of border outline}

\item{auto_recenter}{If TRUE, zooms to area when the border changes (reactively)}

\item{label_function}{Function (with \code{data} and \code{...} as arguments) to make labels. Optional.}

\item{label_params}{List of further arguments to pass to \code{label_function}}

\item{color_default}{Default (fill) color if color is not computed for the layer}

\item{color_outline}{Fixed color of the outline of polygons}

\item{toggle_reload}{Reactive to trigger a reload of the map. Sometimes necessary, such as in modals.}

\item{layers}{List of reactives (each with a list of settings). See example apps.}
}
\description{
Shinto Map server module
}
