% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoMapModule.R
\name{shintoMapUI}
\alias{shintoMapUI}
\title{Shinto Map UI Module}
\usage{
shintoMapUI(id, debugger_panel = FALSE, ...)
}
\arguments{
\item{id}{Shiny ID for the module}

\item{debugger_panel}{If TRUE, add a verbatimtextoutput panel showing all shiny input values, handy for debugging}

\item{\dots}{Further arguments passed to \code{\link[=leafletOutput]{leafletOutput()}}, such as \code{height}, \code{width}}
}
\description{
Shinto Map UI Module
}
\seealso{
\code{\link[=shintoMapModule]{shintoMapModule()}}
}
