% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addshintoMapLayers.R
\name{addShintoMapLayers}
\alias{addShintoMapLayers}
\title{Add map layers}
\usage{
addShintoMapLayers(
  map,
  default = "Esri.WorldGrayCanvas",
  position = "topright"
)
}
\description{
Add map layers
}
