% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_color_functions.R
\name{factor_map_color}
\alias{factor_map_color}
\title{Map colors for a factor variable}
\usage{
factor_map_color(
  vals,
  colors = NULL,
  palette_function = "viridis",
  reverse = FALSE,
  na.color = grDevices::rgb(0, 0, 0, 0),
  bins_predefined = NULL,
  method = c("auto", "predefined"),
  ...
)
}
\description{
Map colors for a factor variable
}
