% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoBaseMap.R
\name{shintoBaseMap}
\alias{shintoBaseMap}
\title{Makes a leaflet map for use in shintoMapModule}
\usage{
shintoBaseMap(set_view = NULL, ...)
}
\description{
Makes a leaflet map for use in shintoMapModule
}
