% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoBaseMap.R
\name{shintoBaseMap}
\alias{shintoBaseMap}
\title{Makes a leaflet map for use in shintoMapModule}
\usage{
shintoBaseMap(
  set_view = NULL,
  fullscreen_option = FALSE,
  layers_button_position = "topright",
  ...
)
}
\arguments{
\item{set_view}{List with components \code{lng}, \code{lat} and \code{zoom} for initial view of the map}

\item{fullscreen_option}{If TRUE, adds a fullscreen button (from leaflet.extras), slightly buggy sometimes}

\item{layers_button_position}{Where to place the button for the layer options (default topright)}

\item{...}{Further arguments passed to \code{\link[=addShintoMapLayers]{addShintoMapLayers()}}}
}
\description{
Makes a leaflet map for use in shintoMapModule
}
