% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoMapModule.R
\name{shintoMapUI}
\alias{shintoMapUI}
\alias{shintoMapModule}
\title{Shinto Map modules}
\usage{
shintoMapUI(id, debugger_panel = FALSE, ...)

shintoMapModule(
  input,
  output,
  session,
  base_map,
  border = shiny::reactive(NULL),
  border_weight = 2,
  border_color = "#4f4f4f",
  auto_recenter = FALSE,
  label_function = function(data, ...) NULL,
  label_params = list(),
  color_default = "#023e8a",
  color_outline = "#FFFFFF",
  toggle_reload = shiny::reactive(TRUE),
  layers = list()
)
}
\description{
Shinto Map modules
}
